cp/ev_cp_e/requirements.txt
# Dependencias para CPs (Engine y Monitor)
kafka-python==2.0.2
cryptography>=41.0.0